requirements-dev.txt
flake8==7.0.0
google-drive-id-extractor-1.2.0
gspread==6.2.1
python-decouple==3.8
requests==2.32.4